requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
defusedxml==0.7.1
Django==4.2
django-cors-headers==4.1.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-filter==23.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
drf-nested-routers==0.93.4
frozenlist==1.3.3
idna==3.4
Markdown==3.4.3
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.4
Pillow==9.5.0
psycopg==3.1.8
psycopg-binary==3.1.8
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2